Add Mox support
This PR does not change the public API at all but will reorganize the primary Mojito base
module functions into a Mojito.Base module meants to be pulled in via `use`.

This PR also includes behaviour and callbacks for all Mojito functions for the purpose
of compatibility with Mox.

The implementation was borrowed from HTTPoison: https://github.com/edgurgel/httpoison/pull/330

Some additional types were added for readability.
